power-manager logic is running in system-server.
Refactoring of system fw. (pm is merged to system-server)
Currently we have two daemon, which are power-manager and system-server.
But we are planing to provide only one daemon, which is for controling all devices on target.
So this patch will merge power-manager into system-server.
Before long, system-server name will be changed 'deviced'.

Change-Id: Ifa9bb4f7a21dc33600cce94b99a68b530d3e813b
Signed-off-by: Krzysztof Sasiak <k.sasiak@samsung.com>