Fix package build for Tizen:IVI
Build of system-server fails for OBS projects with enabled debuginfo and
defined 'simulator' macro with this error:

error: Installed (but unpackaged) file(s) found:
   /usr/lib/debug/.build-id/6e/480d9a978edacf116bd409aa09b97923290f3b
   /usr/lib/debug/.build-id/6e/480d9a978edacf116bd409aa09b97923290f3b.debug
   /usr/lib/debug/usr/bin/restart.debug

The reason of this failure is explained in this bug
https://bugzilla.redhat.com/show_bug.cgi?id=878863#c2:
Debuginfo generation operates on the disk contents and knows exactly
nothing about specs.

In plain English it means that debuginfo doesn't care about %exclude
statements in %files section of the .spec and processes all binaries,
installed into build root.

This fix removes /usr/bin/restart from buildroot to avoid generation of
debug info for it. It also conditionaly includes this binary into the
package only if 'simulator' macro is not defined.

Change-Id: I8368ecc53edb54fb068d394db1abf8d1d73e593a
Signed-off-by: Ed Bartosh <eduard.bartosh@intel.com>